Updating Data(I got more responses!) + Coding
I got a lot more responses on my Google Form so I updated that data and also changed the names of the columns to be more concise.
I also coded for my bar graph and it works but the legend is kind of
useless because its just the labels that are already on the bars (I
might delete it). Additionally, the search_options is almost done now
but there's some errors that keep coming up that I'll have to fix.